Make ContestStat fields byte (from int)

+++ PKHeX.Core/Legality/Encounters/Data/Encounters3.cs
@@ -143,11 +143,11 @@ namespace PKHeX.Core
         private static readonly EncounterStatic3[] Encounter_RSE = ArrayUtil.ConcatAll(Encounter_RSE_Roam, Encounter_RSE_Regular);
         private static readonly EncounterStatic3[] Encounter_FRLG = ArrayUtil.ConcatAll(Encounter_FRLG_Roam, Encounter_FRLG_Stationary);
 
-        private static readonly int[] TradeContest_Cool =   { 30, 05, 05, 05, 05, 10 };
-        private static readonly int[] TradeContest_Beauty = { 05, 30, 05, 05, 05, 10 };
-        private static readonly int[] TradeContest_Cute =   { 05, 05, 30, 05, 05, 10 };
-        private static readonly int[] TradeContest_Clever = { 05, 05, 05, 30, 05, 10 };
-        private static readonly int[] TradeContest_Tough =  { 05, 05, 05, 05, 30, 10 };
+        private static readonly byte[] TradeContest_Cool =   { 30, 05, 05, 05, 05, 10 };
+        private static readonly byte[] TradeContest_Beauty = { 05, 30, 05, 05, 05, 10 };
+        private static readonly byte[] TradeContest_Cute =   { 05, 05, 30, 05, 05, 10 };
+        private static readonly byte[] TradeContest_Clever = { 05, 05, 05, 30, 05, 10 };
+        private static readonly byte[] TradeContest_Tough =  { 05, 05, 05, 05, 30, 10 };
 
         internal static readonly EncounterTrade3[] TradeGift_RSE =
         {
