+++ PKHeX.Core/Legality/Encounters/EncounterStatic/EncounterStatic6.cs
@@ -8,12 +8,12 @@
     {
         public override int Generation => 6;
 
-        public int CNT_Cool   { get; init; }
-        public int CNT_Beauty { get; init; }
-        public int CNT_Cute   { get; init; }
-        public int CNT_Smart  { get; init; }
-        public int CNT_Tough  { get; init; }
-        public int CNT_Sheen  { get; init; }
+        public byte CNT_Cool   { get; init; }
+        public byte CNT_Beauty { get; init; }
+        public byte CNT_Cute   { get; init; }
+        public byte CNT_Smart  { get; init; }
+        public byte CNT_Tough  { get; init; }
+        public byte CNT_Sheen  { get; init; }
 
         protected override bool IsMatchLocation(PKM pkm)
         {
